[refactor] Main.js
메인 리팩토링
