+++ src/pages/Main.js
@@ -3,8 +3,7 @@ import React, { useState, useEffect, useRef, useContext } from "react";
 import Header from "../components/common/Header/Header";
 import styles from "./Main.module.css";
 import { fetchForeCast, fetchForeCast2 } from "../pages/forecast";
-import { SensorDataContext } from "./../API/SensorDataContext";
-import { useLocation, useNavigate } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import API from "../API/api";
 import SensorDetail from "../components/specific/sensorDetail/sensorDetail";
 /*images*/
@@ -12,7 +11,6 @@ import sunIcon from "../assets/images/main/sun_icon.png";
 import cloudyIcon from "../assets/images/main/cloudy_icon.png";
 import semicloudyIcon from "../assets/images/main/semicloudy_icon.png";
 import cloudyrainIcon from "../assets/images/main/cloudyrain_icon.png";
-
 import weatherlocation from "../assets/images/main/location_icon.png";
 import logsensor from "../assets/images/main/sensor_icon.png";
 import currentbuilding from "../assets/images/main/currentBuilding_icon.png";
@@ -262,8 +260,6 @@ function Main() {
   const [hoveredIndex, setHoveredIndex] = useState(null);
   const [sensorData, setSensorData] = useState(null); // 센서 데이터 상태 추가
 
-  const { loading } = useContext(SensorDataContext);
-
   // 좌표 저장
   const coordinates = [
     { building: "신공학관", id: "3115", x: 464, y: 442 },
@@ -335,67 +331,12 @@ function Main() {
     fetchSensorData();
   }, [hoveredIndex]);
 
-  const getTemperatureColor = (value) => {
-    if (value < 16.5 || value > 27.5) return "#F44336";
-    if ((value >= 16.5 && value < 17.6) || (value > 26.4 && value <= 27.5))
-      return "#FF9800";
-    if ((value >= 17.6 && value < 18.7) || (value > 25.3 && value <= 26.4))
-      return "#FFEB3B";
-    if ((value >= 18.7 && value < 19.8) || (value > 24.2 && value <= 25.3))
-      return "#8BC34A";
-    return "#5C82F5";
-  };
-
-  const getHumidityColor = (value) => {
-    if (value < 10 || value > 90) return "#F44336";
-    if ((value >= 10 && value < 20) || (value > 80 && value <= 90))
-      return "#FF9800";
-    if ((value >= 20 && value < 30) || (value > 70 && value <= 80))
-      return "#FFEB3B";
-    if ((value >= 30 && value < 40) || (value > 60 && value <= 70))
-      return "#8BC34A";
-    return "#5C82F5";
-  };
-
-  const getTVOCColor = (value) => {
-    if (value > 10000) return "#F44336";
-    if (value > 3000 && value <= 10000) return "#FF9800";
-    if (value > 1000 && value <= 3000) return "#FFEB3B";
-    if (value > 300 && value <= 1000) return "#8BC34A";
-    return "#5C82F5";
-  };
-
-  const getPM25Color = (value) => {
-    if (value > 64) return "#F44336";
-    if (value > 53 && value <= 64) return "#FF9800";
-    if (value > 41 && value <= 53) return "#FFEB3B";
-    if (value > 23 && value <= 41) return "#8BC34A";
-    return "#5C82F5";
-  };
-
-  const getNoiseColor = (value) => {
-    if (value > 80) return "#F44336";
-    if (value > 70 && value <= 80) return "#FF9800";
-    if (value > 60 && value <= 70) return "#FFEB3B";
-    if (value > 50 && value <= 60) return "#8BC34A";
-    return "#5C82F5";
-  };
-
   const getIAQColor = (value) => {
     if (value > 90) return "#5C82F5";
     if (value > 80 && value <= 90) return "#8BC34A";
     if (value > 70 && value <= 80) return "#FFEB3B";
     if (value > 60 && value <= 70) return "#FF9800";
     return "#F44336";
-  };
-
-  // 센서 타입에 따라 적절한 색상 반환
-  const getStatusColor = (value, type) => {
-    if (type === "temperature") return getTemperatureColor(value);
-    if (type === "humidity") return getHumidityColor(value);
-    if (type === "tvoc") return getTVOCColor(value);
-    if (type === "pm25") return getPM25Color(value);
-    if (type === "noise") return getNoiseColor(value);
   };
 
   const getSensorIAQValue = (id) => {
@@ -471,7 +412,6 @@ function Main() {
               />
               <SensorDetail
                 coord={coordinates[hoveredIndex]}
-                getStatusColor={getStatusColor}
                 hoveredIndex={hoveredIndex}
               />
             </div>
